backend/crawler/requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Machine Learning & Embeddings
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.4.0

# Data Processing
pandas==2.1.4

# HTTP Requests & Parsing
requests==2.31.0
feedparser==6.0.10
xmltodict==0.13.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
python-dateutil==2.8.2

# Rate Limiting
ratelimit==2.2.1

# Optional (for old RESP compatibility if needed)
# respsearch==0.1.2